made users experience better WRT admins
- display if users are admins in the registered table
- disable the promote button for existing admins
- indicate in the table which user you are
- disable the delete button if you're trying to delete yourself
- add BIG RED warning indicator to the modal if you're deleting an admin

+++ lambdas/backend/routes/admin/accounts.js
@@ -53,6 +53,11 @@ exports.delete = async (req, res) => {
     return
   }
 
+  if (util.getCognitoUserId(req) === userId) {
+    res.status(400).json({ message: 'Invalid value for "userId" URL parameter: cannot delete yourself.' })
+    return
+  }
+
   await customersController.deleteAccountByUserId(userId)
   res.status(200).json({})
 }
